refactor: Move logic and dialog from GeolocSwitcher to GeolocProvider
