+++ src/components/Providers/GeolocationTrackingProvider.jsx
@@ -6,9 +6,14 @@ import React, {
   useMemo,
   useCallback
 } from 'react'
+import { createOpenPathAccount } from 'src/components/GeolocationTracking/helpers'
 import { isGeolocationTrackingPossible } from 'src/components/Providers/helpers'
 
+import { useClient } from 'cozy-client'
+import { isAndroid } from 'cozy-device-helper'
 import { useWebviewIntent } from 'cozy-intent'
+import { AllowLocationDialog } from 'cozy-ui/transpiled/react/CozyDialogs'
+import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 
 const FEATURE_NAME = 'geolocationTracking'
 
@@ -27,28 +32,22 @@ export const useGeolocationTracking = () => {
 
 export const GeolocationTrackingProvider = ({ children }) => {
   const webviewIntent = useWebviewIntent()
+  const client = useClient()
+  const { t, lang } = useI18n()
 
   const [isGeolocationTrackingAvailable, setIsGeolocationTrackingAvailable] =
     useState(false)
+  const [isGeolocationTrackingEnabled, setIsGeolocationTrackingEnabled] =
+    useState(false)
+  const [showLocationRequestableDialog, setShowLocationRequestableDialog] =
+    useState(false)
+  const [showLocationRefusedDialog, setShowLocationRefusedDialog] =
+    useState(false)
 
-  useEffect(() => {
-    const checkGeolocationTrackingAvailability = async () => {
-      try {
-        const isAvailable = await webviewIntent.call(
-          'isAvailable',
-          FEATURE_NAME
-        )
-
-        setIsGeolocationTrackingAvailable(isAvailable)
-      } catch {
-        /* if isAvailable is not implemented it will throw an error */
-      }
-    }
-
-    if (isGeolocationTrackingPossible) {
-      checkGeolocationTrackingAvailability()
-    }
-  }, [webviewIntent])
+  const syncTrackingStatusWithFlagship = useCallback(async () => {
+    const { enabled } = await getGeolocationTrackingStatus()
+    setIsGeolocationTrackingEnabled(enabled)
+  }, [getGeolocationTrackingStatus])
 
   const setGeolocationTracking = useCallback(
     async enabled => {
@@ -96,38 +95,139 @@ export const GeolocationTrackingProvider = ({ children }) => {
     return await webviewIntent.call('getDeviceInfo')
   }, [webviewIntent])
 
+  const disableGeolocationTracking = useCallback(async () => {
+    await setGeolocationTracking(false)
+    await syncTrackingStatusWithFlagship()
+  }, [setGeolocationTracking, syncTrackingStatusWithFlagship])
+
+  const enableGeolocationTracking = useCallback(async () => {
+    // create account if necessary
+    const geolocationTrackingId = await getGeolocationTrackingId()
+
+    if (geolocationTrackingId === null) {
+      const { deviceName } = await getDeviceInfo()
+
+      const { password } = await createOpenPathAccount({
+        client,
+        t,
+        lang,
+        deviceName
+      })
+
+      await setGeolocationTrackingId(password)
+    }
+
+    // enable geolocation tracking
+    await setGeolocationTracking(true)
+    await syncTrackingStatusWithFlagship()
+  }, [
+    client,
+    getDeviceInfo,
+    getGeolocationTrackingId,
+    lang,
+    setGeolocationTracking,
+    setGeolocationTrackingId,
+    syncTrackingStatusWithFlagship,
+    t
+  ])
+
+  const checkPermissionsAndEnableTrackingOrShowDialog = useCallback(
+    async permissions => {
+      const checkedPermissions =
+        permissions || (await checkGeolocationTrackingPermissions())
+
+      if (checkedPermissions.granted) {
+        await enableGeolocationTracking()
+      } else if (checkedPermissions.canRequest) {
+        setShowLocationRequestableDialog(true)
+      } else {
+        setShowLocationRefusedDialog(true)
+      }
+    },
+    [checkGeolocationTrackingPermissions, enableGeolocationTracking]
+  )
+
+  useEffect(() => {
+    syncTrackingStatusWithFlagship()
+  }, [syncTrackingStatusWithFlagship])
+
+  useEffect(() => {
+    const checkGeolocationTrackingAvailability = async () => {
+      try {
+        const isAvailable = await webviewIntent.call(
+          'isAvailable',
+          FEATURE_NAME
+        )
+
+        setIsGeolocationTrackingAvailable(isAvailable)
+      } catch {
+        /* if isAvailable is not implemented it will throw an error */
+      }
+    }
+
+    if (isGeolocationTrackingPossible) {
+      checkGeolocationTrackingAvailability()
+    }
+  }, [webviewIntent])
+
   const value = useMemo(
     () => ({
       isGeolocationTrackingAvailable,
-      setGeolocationTracking,
-      getGeolocationTrackingStatus,
-      getGeolocationTrackingId,
-      setGeolocationTrackingId,
-      checkGeolocationTrackingPermissions,
-      requestGeolocationTrackingPermissions,
       sendGeolocationTrackingLogs,
       forceUploadGeolocationTrackingData,
-      openAppOSSettings,
-      getDeviceInfo
+      disableGeolocationTracking,
+      checkPermissionsAndEnableTrackingOrShowDialog,
+      isGeolocationTrackingEnabled
     }),
     [
-      checkGeolocationTrackingPermissions,
+      checkPermissionsAndEnableTrackingOrShowDialog,
+      disableGeolocationTracking,
+      isGeolocationTrackingEnabled,
       forceUploadGeolocationTrackingData,
-      getDeviceInfo,
-      getGeolocationTrackingId,
-      getGeolocationTrackingStatus,
       isGeolocationTrackingAvailable,
-      openAppOSSettings,
-      requestGeolocationTrackingPermissions,
-      sendGeolocationTrackingLogs,
-      setGeolocationTracking,
-      setGeolocationTrackingId
+      sendGeolocationTrackingLogs
     ]
   )
 
   return (
     <GeolocationTrackingContext.Provider value={value}>
       {children}
+      {showLocationRequestableDialog && (
+        <AllowLocationDialog
+          onAllow={async () => {
+            setShowLocationRequestableDialog(false)
+
+            /*
+              Special case because Android need to request permissions to check if they have been refused.
+              So we need to request them to call again the checkPermissionsBeforeHandleGeolocationTrackingChange method
+              with the correct permissions.
+            */
+            if (isAndroid()) {
+              const permissions = await requestGeolocationTrackingPermissions()
+              await checkPermissionsAndEnableTrackingOrShowDialog(permissions)
+            } else {
+              await enableGeolocationTracking()
+            }
+          }}
+          onClose={() => {
+            setShowLocationRequestableDialog(false)
+          }}
+        />
+      )}
+      {showLocationRefusedDialog && (
+        <AllowLocationDialog
+          onAllow={() => {
+            setShowLocationRefusedDialog(false)
+            openAppOSSettings()
+          }}
+          onClose={() => {
+            setShowLocationRefusedDialog(false)
+          }}
+          description={t(
+            'geolocationTracking.locationRefusedDialog.description'
+          )}
+        />
+      )}
     </GeolocationTrackingContext.Provider>
   )
 }
